tests, net: Move service-mesh helpers from net/utils to net/mesh/utils
