+++ tests/network/utils.py
@@ -2,7 +2,6 @@ import logging
 import shlex
 from collections import OrderedDict
 
-import pexpect
 from kubernetes.dynamic.exceptions import ResourceNotFoundError
 from ocp_resources.deployment import Deployment
 from ocp_resources.node_network_state import NodeNetworkState
@@ -10,8 +9,7 @@ from ocp_resources.service import Service
 from pyhelper_utils.shell import run_ssh_commands
 from timeout_sampler import TimeoutExpiredError, TimeoutSampler
 
-from tests.network.constants import BRCNV, SERVICE_MESH_PORT
-from utilities import console
+from tests.network.constants import BRCNV
 from utilities.constants import (
     IPV4_STR,
     OS_FLAVOR_FEDORA,
@@ -272,57 +270,6 @@ def assert_nncp_successfully_configured(nncp):
     except TimeoutExpiredError:
         LOGGER.error(f"{nncp.name} is not {successfully_configured}, but rather {nncp.status}.")
         raise
-
-
-def authentication_request(vm, **kwargs):
-    """
-    Return server response to a request sent from VM console. This request allows testing client authentication.
-
-    Args:
-        vm (VirtualMachine): VM that will be used for console connection
-
-    Kwargs: ( Used to allow passing args from wait_service_mesh_components_convergence in service_mesh/conftest)
-        service (str): target svc dns name
-
-    Returns:
-        str: Server response
-    """
-    return run_console_command(
-        vm=vm,
-        command=f"curl http://{kwargs['service']}:{SERVICE_MESH_PORT}/ip",
-    )
-
-
-def assert_service_mesh_request(expected_output, request_response):
-    assert expected_output in request_response, (
-        f"Server response error.Expected output - {expected_output}received - {request_response}"
-    )
-
-
-def assert_authentication_request(vm, service_app_name):
-    # Envoy proxy IP
-    expected_output = "127.0.0.6"
-    request_response = authentication_request(vm=vm, service=service_app_name)
-    assert_service_mesh_request(expected_output=expected_output, request_response=request_response)
-
-
-def run_console_command(vm, command, timeout=TIMEOUT_1MIN):
-    """
-    Run a single command through a VM console.
-    """
-    prompt = r"\$ "
-    with console.Console(vm=vm, prompt=prompt) as vmc:
-        LOGGER.info(f"Execute {command} on {vm.name}")
-        try:
-            vmc.sendline(command)
-            vmc.expect(prompt, timeout=timeout)
-            return vmc.before
-        except pexpect.exceptions.TIMEOUT:
-            LOGGER.info(f"Timeout: {vmc.before}")
-            return vmc.before
-        except pexpect.exceptions.EOF:
-            LOGGER.info(f"EOF: {vmc.before}")
-            return vmc.before
 
 
 def vm_for_brcnv_tests(
